cambios

+++ Aerolinea/airline-rest-server/src/main/java/org/educa/airline/services/UserService.java
@@ -2,10 +2,13 @@ package org.educa.airline.services;
 
 
 import org.educa.airline.Exceptions.NotExistUser;
+import org.educa.airline.entity.Role;
 import org.educa.airline.entity.User;
 
 import org.educa.airline.repository.inmemory.InMemoryUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -18,7 +21,6 @@ public class UserService implements UserDetailsService {
     private final InMemoryUserRepository inMemoryUserRepository;
 
 
-
     @Autowired
     public UserService(InMemoryUserRepository inMemoryUserRepository) {
         this.inMemoryUserRepository = inMemoryUserRepository;
@@ -26,40 +28,68 @@ public class UserService implements UserDetailsService {
     }
 
 
-    public boolean create(User user) throws NotExistUser {
+    public boolean create(User user) {
         inMemoryUserRepository.createUser(user);
         return true;
     }
 
-    public boolean update(User user, String id, String nif) {
-
-        if (inMemoryUserRepository.existUser(id)) {
-            inMemoryUserRepository.updateUser(user);
-            return true;
+    public boolean update(User user, String username) throws Exception {
+        if (inMemoryUserRepository.existUser(username)) {
+            if (user.getUsername().equals(username)) {
+                inMemoryUserRepository.updateUser(user);
+                return true;
+            } else {
+                if (!inMemoryUserRepository.existUser(user.getUsername())) {
+                    delete(username);
+                    create(user);
+                    return true;
+                } else {
+                    return false;
+                }
+            }
         } else {
             return false;
         }
     }
 
-    public boolean delete(String username, String nif) {
-        return inMemoryUserRepository.deleteUser(username);
+    public boolean delete(String username) throws Exception {
+
+       Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+       UserDetails userDetail = (UserDetails) auth.getPrincipal();
+        if (userDetail.getUsername().equals(username) || userDetail.getAuthorities().contains(new Role("ROLE_admin", "admin", "Es un administrador"))) {
+            if (inMemoryUserRepository.existUser(username)) {
+                inMemoryUserRepository.deleteUser(username);
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+           throw new Exception();
+        }
     }
 
     public List<User> findAllUser() throws NotExistUser {
-            return inMemoryUserRepository.getUsers();
+        return inMemoryUserRepository.getUsers();
     }
 
 
-    public boolean exitPassenger(String username) {
+    public boolean exitUser(String username) {
         return inMemoryUserRepository.existUser(username);
     }
 
-    public User findUser(String email) {
-        return inMemoryUserRepository.getUser(email);
+    public User findUser(String username) {
+        return inMemoryUserRepository.getUser(username);
     }
 
     @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return inMemoryUserRepository.getUser(email);
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return inMemoryUserRepository.getUser(username);
+    }
+
+    public boolean isCurrentUser(String username) {
+        // Obtiene el nombre de usuario actualmente autenticado
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        // Verifica si el usuario autenticado es el mismo que se está intentando eliminar
+        return currentUsername.equals(username);
     }
 }
